extract advisor brief workflow pack adapter
